+++ api-v1-0/deleteUserOrder.php
@@ -8,24 +8,26 @@ if (isset($_POST['userId']) && !empty($_POST['userId'])) {
     $mainDb = $app->getMainDb();
     $channelId = 1; // VK
 
-    try {
-        // $result = $mainDb->delete('user_order',
-        //     ['user_id' => $_POST['userId'], 'id' => $_POST['dbId']],
-        //     ['int', 'int']);
-$result = $mainDb->query('DELETE FROM user_order WHERE id='.$_POST['dbId'].' AND user_id = '.$_POST['userId']);
-        if ($result) {
-            $json_data['message'] = '';
-            echo json_encode($json_data);
-        } else {
-            $json_data['id'] = 2;
-            throw new Exception("Bad request to DB!");
+    if ($app->checkSessionKey($_POST['userId'], $_POST['sessionKey'])) {
+        try {
+            $result = $mainDb->query('DELETE FROM user_order WHERE id='.$_POST['dbId'].' AND user_id = '.$_POST['userId']);
+            if ($result) {
+                $json_data['message'] = '';
+                echo json_encode($json_data);
+            } else {
+                $json_data['id'] = 2;
+                throw new Exception("Bad request to DB!");
+            }
         }
-    }
-
-    catch (Exception $e)
-    {
-        $json_data['status'] = 's058';
-        $json_data['message'] = $e->getMessage();
+        catch (Exception $e)
+        {
+            $json_data['status'] = 's058';
+            $json_data['message'] = $e->getMessage();
+            echo json_encode($json_data);
+        }
+    } else {
+        $json_data['id'] = 13;
+        $json_data['message'] = 'bad sessionKey';
         echo json_encode($json_data);
     }
 }
