+++ api-v1-0/releaseUserTrainPack.php
@@ -8,28 +8,34 @@ if (isset($_POST['userId']) && !empty($_POST['userId'])) {
     $mainDb = $app->getMainDb();
     $channelId = 1; // VK
 
-    try {
-        // $result = $mainDb->delete('user_train_pack',
-        //     ['user_id' => $_POST['userId']],
-        //     ['int']);
-        $result = $mainDb->query('DELETE FROM user_train_pack WHERE user_id='.$_POST['userId']);
-        // $result2 = $mainDb->delete('user_train_pack_item',
-        //     ['user_id' => $_POST['userId']],
-        //     ['int']);
-        $result2 = $mainDb->query('DELETE FROM user_train_pack_item WHERE user_train_pack_id='.$_POST['id']);
-        if ($result && $result2) {
-            $json_data['message'] = '';
-            echo json_encode($json_data);
-        } else {
-            $json_data['id'] = 2;
-            throw new Exception("Bad request to DB!");
+    if ($app->checkSessionKey($_POST['userId'], $_POST['sessionKey'])) {
+        try {
+            // $result = $mainDb->delete('user_train_pack',
+            //     ['user_id' => $_POST['userId']],
+            //     ['int']);
+            $result = $mainDb->query('DELETE FROM user_train_pack WHERE user_id='.$_POST['userId']);
+            // $result2 = $mainDb->delete('user_train_pack_item',
+            //     ['user_id' => $_POST['userId']],
+            //     ['int']);
+            $result2 = $mainDb->query('DELETE FROM user_train_pack_item WHERE user_train_pack_id='.$_POST['id']);
+            if ($result && $result2) {
+                $json_data['message'] = '';
+                echo json_encode($json_data);
+            } else {
+                $json_data['id'] = 2;
+                throw new Exception("Bad request to DB!");
+            }
         }
-    }
 
-    catch (Exception $e)
-    {
-        $json_data['status'] = 's142';
-        $json_data['message'] = $e->getMessage();
+        catch (Exception $e)
+        {
+            $json_data['status'] = 's142';
+            $json_data['message'] = $e->getMessage();
+            echo json_encode($json_data);
+        }
+    } else {
+        $json_data['id'] = 13;
+        $json_data['message'] = 'bad sessionKey';
         echo json_encode($json_data);
     }
 }
